Return an empty response instead of throw an error
There's a buggy version of Mono (currently used by Unity 2020) that doesn't seem to return this stream correct.
It looks like the issue occurs when there's no content length in an error response.
Unfortunately we can't get the body of the response when this happens, but we can at least not throw here and return the proper error code